requirements.txt
openai
chromadb
langchain-community
langchain-text-splitters
pypdf
streamlit
python-dotenv